requirements.txt
tweepy>=4.14.0
anthropic>=0.18.0
python-dotenv>=1.0.0
